[UPDATE]: link frontend sign pages to backend + handle errors
